Added CORS enabling method
Added cross origin resource shraring(CORS) enabling method, and
kept the header `Access-Control-Allow-Origin: *`.This can result
in an attacker possibly placing a proxy between a victim and the
CORS enabled site, thus allowing the attacker to inject any script
to the victims site. Thus when enabling CORS for a publicly
accessible site, make sure the `Access-Control-Allow-Origin` is a
controlled list of whitelisted URIs.